Move conversion "from Ctf event to EVENT_RECORD" in its own class
